requirements.txt
playwright==1.59.0
pytest==9.0.3
pytest-html==4.2.0
Faker==40.15.0
python-dotenv==1.2.2
